Adding extension to enrich the prompt & refactor
