+++ src/cli.ts
@@ -28,7 +28,9 @@ process.env.PI_SKIP_VERSION_CHECK = "1"
 import { loadConfig } from "./config.js"
 import subagentExtension from "./extensions/subagent.js"
 import webFetchExtension from "./extensions/web-fetch/index.js"
+import promptEnrichmentExtension from "./extensions/prompt-enrichment/index.js"
 import { updateModelsConfig } from "./models.js"
+
 
 try {
 	const config = loadConfig()
