+++ src/orchestration/model-registry/index.ts
@@ -1,4 +1,3 @@
 export { ModelRegistry } from "./model-registry.js"
 export { BUILTIN_MODELS } from "./builtin-models.js"
-export { transformPrompt } from "./prompt-transformer.js"
 export type { ModelStrength, ModelCapabilities, OrchestrationModelDescriptor } from "./types.js"
